requirements.txt
azure-identity>=1.15.0
msgraph-core>=1.0.0
msgraph-sdk
PyYAML>=6.0.1
openai>=1.12.0
anthropic>=0.18.1
colorama>=0.4.6
fastapi>=0.104.1
uvicorn>=0.24.0
websockets>=12.0